- Move pre/postprocessing up a directory - Add node2vec postprocessing - Save .pkl embeddings by concept
